update index.ts in bb backend interface file

+++ tooling/noir_js_backend_barretenberg/src/index.ts
@@ -1,21 +1,6 @@
 /* eslint-disable  @typescript-eslint/no-explicit-any */
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-
-// import { Barretenberg, Crs, RawBuffer } from '@aztec/bb.js/dest/browser/types/index.js';
-import { decompressSync as gunzip } from 'fflate';
-import { Backend, CompiledCircuit } from '@noir-lang/types';
-
-// Since this is a simple function, we can use feature detection to
-// see if we are in the nodeJs environment or the browser environment.
-export function base64Decode(input: string): Uint8Array {
-  return Uint8Array.from(atob(input), (c) => c.charCodeAt(0));
-}
-
-// Converts an bytecode to a Uint8Array
-export function acirToUint8Array(base64EncodedBytecode): Uint8Array {
-  const compressedByteCode = base64Decode(base64EncodedBytecode);
-  return gunzip(compressedByteCode);
-}
+import { acirToUint8Array } from "./serialize.js";
+import { Backend, CompiledCircuit } from "@noir-lang/types";
 
 export class BarretenbergBackend implements Backend {
   // These type assertions are used so that we don't
@@ -33,17 +18,23 @@ export class BarretenbergBackend implements Backend {
     this.acirUncompressedBytecode = acirToUint8Array(acirBytecodeBase64);
   }
 
-  async instantiate(): Promise<void> {
+  private async instantiate(): Promise<void> {
     if (!this.api) {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       //@ts-ignore
-      const { Barretenberg, RawBuffer, Crs } = await import('@aztec/bb.js');
+      const { Barretenberg, RawBuffer, Crs } = await import("@aztec/bb.js");
       const api = await Barretenberg.new(this.numberOfThreads);
 
-      const [_exact, _total, subgroupSize] = await api.acirGetCircuitSizes(this.acirUncompressedBytecode);
+      const [_exact, _total, subgroupSize] = await api.acirGetCircuitSizes(
+        this.acirUncompressedBytecode,
+      );
       const crs = await Crs.new(subgroupSize + 1);
       await api.commonInitSlabAllocator(subgroupSize);
-      await api.srsInitSrs(new RawBuffer(crs.getG1Data()), crs.numPoints, new RawBuffer(crs.getG2Data()));
+      await api.srsInitSrs(
+        new RawBuffer(crs.getG1Data()),
+        crs.numPoints,
+        new RawBuffer(crs.getG2Data()),
+      );
 
       this.acirComposer = await api.acirNewAcirComposer(subgroupSize);
       this.api = api;
@@ -55,16 +46,11 @@ export class BarretenbergBackend implements Backend {
   //
   // The settings for this proof are the same as the settings for a "normal" proof
   // ie one that is not in the recursive setting.
-  async generateProof(decompressedWitness: Uint8Array, optimizeForVerifyInCircuit = false): Promise<Uint8Array> {
-    await this.instantiate();
-    const proof = await this.api.acirCreateProof(
-      this.acirComposer,
-      this.acirUncompressedBytecode,
-      decompressedWitness,
-      optimizeForVerifyInCircuit,
-    );
-
-    return proof;
+  async generateFinalProof(
+    decompressedWitness: Uint8Array,
+  ): Promise<Uint8Array> {
+    const makeEasyToVerifyInCircuit = false;
+    return this.generateProof(decompressedWitness, makeEasyToVerifyInCircuit);
   }
 
   // Generates an inner proof. This is the proof that will be verified
@@ -78,9 +64,24 @@ export class BarretenbergBackend implements Backend {
   // We set `makeEasyToVerifyInCircuit` to true, which will tell the backend to
   // generate the proof using components that will make the proof
   // easier to verify in a circuit.
-  async generateChildProof(witness: Uint8Array): Promise<Uint8Array> {
-    const optimizeForVerifyInCircuit = true;
-    return this.generateProof(witness, optimizeForVerifyInCircuit);
+  async generateIntermediateProof(witness: Uint8Array): Promise<Uint8Array> {
+    const makeEasyToVerifyInCircuit = true;
+    return this.generateProof(witness, makeEasyToVerifyInCircuit);
+  }
+
+  async generateProof(
+    decompressedWitness: Uint8Array,
+    makeEasyToVerifyInCircuit: boolean,
+  ): Promise<Uint8Array> {
+    await this.instantiate();
+    const proof = await this.api.acirCreateProof(
+      this.acirComposer,
+      this.acirUncompressedBytecode,
+      decompressedWitness,
+      makeEasyToVerifyInCircuit,
+    );
+
+    return proof;
   }
 
   // Generates artifacts that will be passed to a circuit that will verify this proof.
@@ -92,7 +93,7 @@ export class BarretenbergBackend implements Backend {
   // method.
   //
   // The number of public inputs denotes how many public inputs are in the inner proof.
-  async generateChildProofArtifacts(
+  async generateIntermediateProofArtifacts(
     proof: Uint8Array,
     numOfPublicInputs = 0,
   ): Promise<{
@@ -101,14 +102,20 @@ export class BarretenbergBackend implements Backend {
     vkHash: string;
   }> {
     await this.instantiate();
-    const proofAsFields = await this.api.acirSerializeProofIntoFields(this.acirComposer, proof, numOfPublicInputs);
+    const proofAsFields = await this.api.acirSerializeProofIntoFields(
+      this.acirComposer,
+      proof,
+      numOfPublicInputs,
+    );
 
     // TODO: perhaps we should put this in the init function. Need to benchmark
     // TODO how long it takes.
     await this.api.acirInitVerificationKey(this.acirComposer);
 
     // Note: If you don't init verification key, `acirSerializeVerificationKeyIntoFields`` will just hang on serialization
-    const vk = await this.api.acirSerializeVerificationKeyIntoFields(this.acirComposer);
+    const vk = await this.api.acirSerializeVerificationKeyIntoFields(
+      this.acirComposer,
+    );
 
     return {
       proofAsFields: proofAsFields.map((p) => p.toString()),
@@ -117,18 +124,34 @@ export class BarretenbergBackend implements Backend {
     };
   }
 
-  async verifyChildProof(proof: Uint8Array): Promise<boolean> {
-    const optimizeForVerifyInCircuit = true;
-    return this.verifyProof(proof, optimizeForVerifyInCircuit);
+  async verifyFinalProof(proof: Uint8Array): Promise<boolean> {
+    const makeEasyToVerifyInCircuit = false;
+    const verified = await this.verifyProof(proof, makeEasyToVerifyInCircuit);
+    return verified;
   }
 
-  async verifyProof(proof: Uint8Array, optimizeForVerifyInCircuit = false): Promise<boolean> {
+  async verifyIntermediateProof(proof: Uint8Array): Promise<boolean> {
+    const makeEasyToVerifyInCircuit = true;
+    return this.verifyProof(proof, makeEasyToVerifyInCircuit);
+  }
+
+  async verifyProof(
+    proof: Uint8Array,
+    makeEasyToVerifyInCircuit: boolean,
+  ): Promise<boolean> {
     await this.instantiate();
     await this.api.acirInitVerificationKey(this.acirComposer);
-    return await this.api.acirVerifyProof(this.acirComposer, proof, optimizeForVerifyInCircuit);
+    return await this.api.acirVerifyProof(
+      this.acirComposer,
+      proof,
+      makeEasyToVerifyInCircuit,
+    );
   }
 
   async destroy(): Promise<void> {
+    if (!this.api) {
+      return;
+    }
     await this.api.destroy();
   }
 }
